// src/app/models/Cars.Model.ts
export interface Car {
  carId: number;
  carName: string;
  carImage: string;
  carBrand: string;
  carFuelType: string;
  carTransmissionType: string;
  carIsHybrid: boolean;
  carIsNew: boolean;
  carSegment: string;
  cityId: number;
  cityName: string;
  carRentalId: number;
  carAmount: number;
  carTax: number;
  carDiscount: number;
  carAvailableIn: number;
}